requirements.txt
discord.py==1.6.0
discord==1.0.1
python-dotenv==0.16.0
youtube-dl==2021.6.6
PyNaCl==1.4.0
asyncio==3.4.3
pafy==0.5.5
dismusic
wavelink
discord-custom-help
